requirements.dev.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
exceptiongroup==1.1.1
fonttools==4.40.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
kiwisolver==1.4.4
lz4==4.3.2
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.25.0
opensafely-cohort-extractor==1.84.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
py==1.11.0
pyarrow==12.0.1
pyparsing==3.1.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
retry==0.9.2
scipy==1.11.0
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
structlog==23.1.0
tabulate==0.9.0
tomli==2.0.1
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zipp==3.15.0
